moved the default avatar based transform picker to a seperate hook to be used in different places, also applied it to all current avatars and it seems to work nice
